requirements.txt
python-telegram-bot==20.3
aiohttp==3.9.1
openai==1.3.7
loguru>=0.7.0
apscheduler>=3.11.0 